add session mangagement. issue #25
Session is client side signed cookie session.
When a user requests a page with a query param of "source"
all of the passed query params are added to the session. This
facilitates tracking conversion from marketing campaigns.
When a user submits the signup form, all tracking data is passed
to intercom with "tracking_" prepented to the query param name.